Support domain verification
SimpleLogin now requires a verification code be posted as a TXT record
before you can use a custom domain. However, domains added before that
requirement was introduced do not retroactively get such a code, which
is why the corresponding variable is optional and the record is not
created if it is not defined.